Finished the merge, introduced the compression write and the non comp…

+++ inverted-index/src/main/java/it/unipi/mircv/beans/Posting.java
@@ -1,15 +1,15 @@
 package it.unipi.mircv.beans;
 
 public class Posting {
-    Integer doc_id;
+    long doc_id;
     Integer frequency;
 
-    public Posting(Integer doc_id, Integer frequency) {
+    public Posting(long doc_id, Integer frequency) {
         this.doc_id = doc_id;
         this.frequency = frequency;
     }
 
-    public Integer getDoc_id() {
+    public long getDoc_id() {
         return doc_id;
     }
